requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
pydantic[email]>=2.5.3

# File handling
aiofiles>=23.2.1

# Optional: YAML support for config validator
pyyaml>=6.0.1

# Development dependencies (optional, install separately)
# pytest>=7.4.4
# pytest-asyncio>=0.23.3
# httpx>=0.26.0
# black>=24.1.1
# ruff>=0.1.14
